Stop entities -- Javadoc
-move LocationType enum in the class that actually use it, making it private at the same occasion
