+++ domain/src/main/java/org/mobilitydata/gtfsvalidator/domain/entity/stops/LocationBase.java
@@ -19,34 +19,10 @@ package org.mobilitydata.gtfsvalidator.domain.entity.stops;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.util.stream.Stream;
-
+/**
+ * Base class for all entities defined in stops.txt
+ */
 public abstract class LocationBase {
-
-    // see https://gtfs.org/reference/static#stopstxt
-    public enum LocationType {
-        STOP_OR_PLATFORM(0),
-        STATION(1),
-        ENTRANCE_OR_EXIT(2),
-        GENERIC_NODE(3),
-        BOARDING_AREA(4);
-
-        private int value;
-
-        LocationType(int value) {
-            this.value = value;
-        }
-
-        static public LocationType fromInt(Integer fromValue) {
-            if (fromValue == null) {
-                return STOP_OR_PLATFORM;
-            }
-            return Stream.of(LocationType.values())
-                    .filter(enumItem -> enumItem.value == fromValue)
-                    .findAny()
-                    .orElse(STOP_OR_PLATFORM);
-        }
-    }
 
     @NotNull
     public String getStopId() {
